python/week03/backend/product_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
azure-storage-blob
python-dotenv